Finish parsing
