+++ server/run.go
@@ -1,17 +1,19 @@
 package main
 
 import (
-    log "github.com/Sirupsen/logrus"
+    "github.com/Sirupsen/logrus"
     "./websocket"
 )
-
-var port = ":3000"
+var (
+    port = ":3000"
+    log = logrus.New()
+)
 
 func main() {
     log.Formatter = new(logrus.TextFormatter)
     log.Level = logrus.DebugLevel
-    
-    log.Info("Attempting to bind socket.")
+    log.Info("Logger initialized.")
 
+    log.Info("Attempting to bind socket.")
     websocket.Listen(port)
 }
